requirements.txt
Flask~=2.1.1
pandas~=1.4.2
requests~=2.27.1
PyYAML~=6.0